DRONE 80% Done
That's right. The drone's movement system is working perfectly as of right now. It should be very plug and play right now as you just need to put the prefab in along with using the Camera prefab in this package as well.

Oh and I pushed a lot of my placeholders and other Test objects into prefab so that it'll be easier to port them over to the main scene when that is done.

Next up will be trying to get the minimap system to work properly. Which means having something that keeps watch of if you are clicking on something, and if you do, select that item. This will likely be rather difficult, but I'm hoping to have it done by, if not on, Sunday.